requirements.txt
PyQt6>=6.8.0
pyotp>=2.9.0
pyzbar>=0.1.9
mss>=10.1.0
Pillow>=11.1.0
pyinstaller>=6.12.0
